Add gitignore to repo
This is usually the first thing I like to do in a repo. Adding the
ignore file is a good first step; it sets you up to not commit anything
by accident and it makes a precedence for smaller unit sized commits.

Hopefully this will be kept up when developing the codebase.

I copied this ignore file from here:

https://github.com/github/gitignore/blob/e58b3e0a015871e4caafd2eed04dc243ef5f30b4/Node.gitignore

Feel free to add things as needed.