Add build step that allows for TypeScript use
=================================================
Add package.json, allowing TypesSript development
=================================================

I really like using TypeSciprt. If you fully type things, then the
process of developing features begins with defining types, which forces
you to think about what you are buliding.

What I really don't like is thinking about the build step, so I opted to
use Parcel. It makes it really simple to bundle files.

Note that the main file is `index.html`. In here, we can reference a
TypeScript file using the `script` tag, and Parcel will figure out that
it needs to convert it to JavaScript. Pretty nifty.

Notice that the build step outputs to a `public` directory, and not the
Parcel default.

::

  +{
  +  "name": "before-after",
  +  "version": "1.0.0",
  +  "description": "",
  +  "scripts": {
  +    "start": "parcel index.html",
  +    "build": "parcel build -d public index.html"
  +  },
  +  "repository": {
  +    "type": "git",
  +    "url": "git+https://github.com/isthisnagee/before-after.git"
  +  },
  +  "keywords": [],
  +  "author": "",
  +  "license": "ISC",
  +  "bugs": {
  +    "url": "https://github.com/isthisnagee/before-after/issues"
  +  },
  +  "homepage": "https://github.com/isthisnagee/before-after#readme",
  +  "dependencies": {},
  +  "devDependencies": {
  +    "parcel-bundler": "^1.12.3",
  +    "typescript": "^3.5.3"
  +  }
  +}

======================================
Create simple HTML and TypeScript file
======================================

Just to test that everything works, I'll create a simple HTML and
TypeScript file, then run `yarn start` and see if it alerts me.

::

  +<!DOCTYPE html>
  +<html lang="en">
  +  <head>
  +    <meta charset="utf-8" />
  +    <meta name="viewport" content="width=device-width" />
  +    <title>before v after</title>
  +  </head>
  +  <body>
  +    <h1>before v after</h1>
  +    <script src="./src/index.ts"></script>
  +  </body>
  +</html>

::

  +alert('hello world');

===================
Ignore build output
===================

This step is not absolutelty necessary at the moment, since the
generated `public` directory could be public. But I think that whatever
machine is hosting this code generate the files.

::

  +# output directory (see package.json)
  +public/
  +

To ignore the output of parcel during development::

  +# ignore parcel build directory
  +dist/
  +
  +