Describe file structure in index.ts file
I feel like having comments in the file describing the structure of
things is incredibly useful, but also hard to maintain.

Keeping it local to the file means that it is hard for developers to
ignore. But it also means that when changing how things work, developers
have to update it. If this update step is ignored, having all that text
there will just create confusion for new developers.

Keeping the comments up to date is something that needs to be watched
out for during code reviews.