Add styling to index.html
The first thing we'll do is add a link to a stylesheet, then we'll add
some classes to the HTML. We'll decide from here that selectors should
be Ids, and styling should happen via classes _only_::

  diff --git a/index.html b/index.html
  index e3814c1..3f22be4 100644
  --- a/index.html
  +++ b/index.html
  @@ -3,10 +3,18 @@
     <head>
       <meta charset="utf-8" />
       <meta name="viewport" content="width=device-width" />
  +    <link rel="stylesheet" href="style.css" type="text/css" media="all" />
       <title>before v after</title>
     </head>
     <body>
  -    <h1>before v after</h1>
  +    <div class="wrapper">
  +      <div class="sidebar">
  +        <div class="sidebar-content">
  +          <h1>before<br />v<br />after</h1>
  +        </div>
  +      </div>
  +      <div class="content"></div>
  +    </div>
       <script src="./src/index.ts"></script>
     </body>
   </html>

Currently the TypeScript file sends an `alert`, which is extremely
annoying during development. Every time the page reloads we have to
clear the alert message. Because we're going to start doing style
changes, removing the alert will allow us to work faster::

  diff --git a/src/index.ts b/src/index.ts
  index 9dcbf0c..a67349e 100644
  --- a/src/index.ts
  +++ b/src/index.ts
  @@ -42,6 +42,6 @@
     main();

     function main(): void {
  -    alert(TEXT.greeting);
  +    console.log(TEXT.greeting);
     }
   })();

Finally, we'll add a stylesheet. Notice that the media query overrides
are for the desktop and not for the mobile. Though this is a desktop
first site, building with mobile in mind is not difficult and it is good
practice. Arguably, knowing when to ignore rules is an important skill
to have, but most interenet traffic happens on phones, and you cannot
understand user needs before talking to users::

  diff --git a/style.css b/style.css
  new file mode 100644
  index 0000000..82fd941
  --- /dev/null
  +++ b/style.css
  @@ -0,0 +1,43 @@
  +* {
  +  margin: 0;
  +  padding: 0;
  +  font-size: 16px;
  +  font-family: sans-serif;
  +}
  +
  +.wrapper {
  +  display: flex;
  +  flex-direction: column;
  +
  +  height: 100vh;
  +}
  +
  +.sidebar {
  +  width: 100%;
  +
  +  background-color: snow;
  +}
  +
  +.sidebar-content {
  +  padding: 0.5rem;
  +}
  +
  +.content {
  +  height: 100%;
  +  width: 100%;
  +
  +  background-color: white;
  +
  +  box-shadow: 3px 3px 5px 6px #ccc;
  +}
  +
  +@media screen and (min-width: 30em) {
  +  .wrapper {
  +    flex-direction: row;
  +  }
  +
  +  .sidebar {
  +    width: 300px;
  +    height: 100%;
  +  }
  +}