Add example of handling input
We'll first create an input element in the HTML.
Notice that we add a label. Always add a label::

  +          <div>
  +            <label for="general--file-name">File Name</label>
  +            <input
  +              type="text"
  +              name="general--file-name"
  +              id="general--file-name"
  +              value=""
  +            />
  +          </div>

Now we want to add some handlers in the JavaScript, but first we'll
create some types to represent the state of the program::

  +  // *****
  +  // TYPES
  +  // *****
  +  //
  +  // This section will contain all shared types.
  +
  +  type Maybe<T> = T | null;
  +
  +  type FormValue = {
  +    general: {
  +      fileName: Maybe<string>;
  +    };
  +  };
  +

Our last commit (ccb610a34c5) added a strict tsconfig file, which means
that if we think a variable can be null, we should say it can be null.
Some TypeScript guides use `?` to signify nullability, for example::

   type FormValue = {
     general: {
       fileName?: string;
     }
   }

But then this means that the following is a valid `FormValue`::

  const formValue: FormValue = {
    general: {}
  }

While in practice they are almost equivalent, I do want to be able to
write::

  const formValue: FormValue = {
    general: { fileName: null }
  }

Now that we have types, we can define some values::

     const TEXT = {
  -    greeting: 'hello world',
  +    initial: {fileName: 'before-v-after'},
     };

  +  const FORM: FormValue = {
  +    general: {
  +      fileName: null,
  +    },
  +  };
  +
  +  const $ = {
  +    general: {
  +      fileName: document.getElementById(
  +        'general--file-name',
  +      ) as HTMLInputElement,
  +    },
  +  };

We use `$` as a shorthand because it's a single character and it looks
different than the typical `[a-zA-Z0-9]+` names. Now that we have some
HTML elements, let's add some handlers::

     function main(): void {
  -    console.log(TEXT.greeting);
  +    $.general.fileName.value = TEXT.initial.fileName;
  +    $.general.fileName.onchange = handleFileNameChange;
  +  }

Keeping with the convention of the file, we'll define functions below
their usage. Here's the handler::

  +  // ********
  +  // HANDLERS
  +  // ********
  +  function handleFileNameChange(e: Event): void {
  +    const currentValue = (e.currentTarget as HTMLInputElement).value;
  +    FORM.general.fileName = currentValue;
     }