Add submit button
Adding a submit button will help us check that the setting the form
variables works as expected. We'll start off by adding the HTML::

  +          <div>
  +            <button id="submit" type="submit">Submit</button>(prints to console)
  +          </div>

Now we'll go and add the handlers for this, but first we'll do some
cleanup::

  -  type Maybe<T> = T | null;
  -
     type FormValue = {
       general: {
  -      fileName: Maybe<string>;
  +      fileName: string;
       };
     };

     const FORM: FormValue = {
       general: {
  -      fileName: null,
  +      fileName: TEXT.initial.fileName,
       },
     };

By making the initial value a string, we no longer need the `Maybe`
type. We can add it back later if we need it.

Following the pattern of the file, we'll first add the submit button to
the `$` const::

  +    submit: document.getElementById('submit') as HTMLButtonElement,

Then we'll go ahead and update the main and add a handler

   function main(): void {
-    $.general.fileName.value = TEXT.initial.fileName;
+    $.general.fileName.value = FORM.general.fileName;
     $.general.fileName.onchange = handleFileNameChange;
+
+    $.submit.onclick = handleSubmit;
   }

+  function handleSubmit(): void {
+    console.log(FORM);
+  }